requirements.txt
# Core dependencies
Flask
flask-cors
Pillow

# Data science libraries
numpy<2.0
pandas
matplotlib
seaborn
opencv-python
scikit-learn
tqdm

# TensorFlow wheels are Python-version specific and are not available for Python 3.13/3.14.
# Use Python 3.11 (recommended) or 3.12 for TensorFlow work.
tensorflow==2.15.1; python_version < "3.12"
tensorflow==2.16.1; python_version >= "3.12" and python_version < "3.13"
